ImageEditor .jsx => .tsx
